+++ src/core/remote/index.ts
@@ -1,51 +1,3 @@
-import { XjsTypes } from '../xjs/types';
-import request from './request';
-import { RequestResult, RemoteRequest, ProxyRequest } from './types';
-import { parse, stringify } from '../../helpers';
-
-export default class Remote {
-  private sender: Function;
-
-  private type: XjsTypes;
-  private exec: Function;
-
-  constructor({ type, exec }) {
-    this.type = type;
-    this.exec = exec;
-  }
-
-  setSender(sender) {
-    this.sender = sender;
-  }
-
-  // used only by remote
-  send(message: RemoteRequest): Promise<any> {
-    return request.register(message, this.sender);
-  }
-
-  receiveMessage(message) {
-    const result = parse(message);
-
-    if (this.isRemote()) {
-      return request.runCallback(result as RequestResult);
-    }
-
-    // proxy
-    this.processRequest(result);
-  }
-
-  async processRequest({ asyncId, fn, args }: ProxyRequest) {
-    const result = await this.exec(fn, ...args);
-
-    this.sender(
-      stringify({
-        asyncId,
-        result,
-      })
-    );
-  }
-
-  private isRemote() {
-    return this.type === XjsTypes.Remote;
-  }
-}
+export { default } from './remote';
+export { default as request } from './request';
+export * from './types';
